LIBHYDRA-564. Moved SendJsonRequest out of Plastron::HTTP module
Moved the "app/services/plastron_services/http/send_json_request.rb"
file to "app/services/send_json_request.rb" as it is a generic class
that is not specific to Plastron, and is useful for other services.

Updated "Plastron::HTTP::SendActivityStreamMessage" class to reflect
the change.

https://umd-dit.atlassian.net/browse/LIBHYDRA-564
